functions/cron/requirements.txt
requests==2.31.0
python-decouple==3.7
boto3==1.26.153